Remove a number of getAuthzSubjectValue() calls.  Changed ResourceEvent to store a subject ID and then look it up later.

+++ src/org/hyperic/hq/appdef/server/session/ApplicationManagerEJBImpl.java
@@ -476,7 +476,7 @@ public class ApplicationManagerEJBImpl extends AppdefSessionEJB
                 AppdefEntityID anId = null;
                 if(appSvc.isIsGroup()) {
                     ResourceGroup group = appSvc.getResourceGroup();
-                    anId = AppdefEntityID.newGroupID(group.getId().intValue());
+                    anId = AppdefEntityID.newGroupID(group.getId());
                 } else {
                     anId = AppdefEntityID.newServiceID(appSvc.getService()
                                                            .getId().intValue());
